Update README.md with installation instructions
Hi,

Love this project- it's one of those perfect packages that does one thing perfectly, and I use it almost every day! 

I've shared this with a few other people, and keep getting e-mails back asking how to install it. While the pip link is easy enough to find, it might be easier for new users to explicitly mention how it can be added to a Python environment, so I added sample code, a link, and the repo name in pip.

Hope this fits within this project, and I'm more than happy to adjust the wording here.