changing dependency requirements
Caret notation in poetry allows for a range but the range is too small. For example, for
`pandas ="^1.5.3"` this would allow for >=1.5.3 <2.0.0 but the current version is 2.2.2. Poetry also
allows for the <= notation so this seems like a better solution, which is why I went with it in my PR.